Fix dialog and display demo folder structure
