+++ .eslintignore
@@ -3,4 +3,3 @@ www
 node_modules
 src/components.d.ts
 src/model/Notice.ts
-src/components/display-demo/index.tsx
